+++ src/Aerofly/MissionCheckpoint.ts
@@ -237,7 +237,7 @@ export class MissionCheckpoint {
     this.slope = altDifference / this.distance_m;
   }
 
-  toString(index: number): string {
+  toString(index: number, noGuides: boolean = false): string {
     let flyOver = "";
     if (this.type === MissionCheckpoint.TYPE_WAYPOINT) {
       flyOver = `\
@@ -249,7 +249,7 @@ export class MissionCheckpoint {
                         <[string8u][type][${Quote.tmc(this.type)}]>
                         <[string8u][name][${Quote.tmc(this.name)}]>
                         <[vector2_float64][lon_lat][${this.lon_lat}]>
-                        <[float64][altitude][${this.lon_lat.altitude_m}]>
+                        <[float64][altitude][${noGuides ? -100 : this.lon_lat.altitude_m}]>
                         //<[float64][speed][${this.speed}]>
                         <[float64][direction][${this.direction}]>
                         <[float64][slope][${this.slope}]> // ${this.slope_deg.toFixed(1)} deg